+++ src/timer.h
@@ -8,11 +8,11 @@
 
 class Timer {
 public:
-    Timer(int ms);
-    int time_remaining();
-    int time_elapsed();
+    Timer(int64_t ms);
+    int64_t time_remaining();
+    int64_t time_elapsed();
 private:
-    int duration; // Total duration of the timer
+    int64_t duration; // Total duration of the timer
     std::chrono::high_resolution_clock::time_point start_time; // Timer start time
 };
